Further enhancements for German browsers
In German browser, tweets with more than 1000 replies/retweet/favorites are displayed as "1,1Tsd. " instead of "1.1K".
The previous version already converted "K" into 1000. Now, the program is able to convert "Tsd. " into 1000 as well.
I also added a function that is able to convert commas into dots. Now, the program is able to convert "1,1" into "1.1" if identified as number of replies/retweets/favorites and thus identify the number as a float.